reuse gRPC connection
